refactor(frontend): decompose listen view into region sub-components (plan 60)
Behaviour-neutral lift. `src/views/listen.tsx` (1136 → 319 lines) is now a slice-wiring orchestrator that composes three new region sub-components under `src/components/listen/`: `listen-header.tsx` (cover + title + book-meta + cover Replace/Regenerate — header section + bottom metadata editor), `listen-player-region.tsx` (markers panel from plan 53 + capped chapter list driving the global mini-player), and `listen-download-section.tsx` (listener-app tiles + export queue + the three "Or download a file" tiles).

Zero spec modifications. All 1046 frontend Vitest tests, 36 listen-specific tests, 13 listen-touching e2e tests (listen-playback, listen-resume, mini-player-features, download-tiles, cover-framing), and the full `npm run verify` battery (lint + typecheck + test:hooks + test + test:server + test:scripts + test:sidecar + test:e2e + build) green on the new shape without touching any test file — that's what verifies the lift is behaviour-neutral.

Gates Wave 3 of the v1.4.0 alpha-launch slate (streaming-link tile, editorial notes, share-clip): each future feature now mounts JSX inside the matching region sub-component instead of injecting into a 1000+ line monolith, so the next three parallel branches no longer collide on the same hot file.

Regression plan: docs/features/archive/60-listen-view-decompose.md.
