+++ docs/features/INDEX.md
@@ -86,6 +86,7 @@ When a plan reaches **stable** AND has a filled **Ship notes** section, move it 
 - [47 — Listening progress / resume bookmarks](archive/47-listen-progress.md) — Per-book sibling `listen-progress.json`; mini-player seeks to the saved point on chapter mount; Listen view "Resume at MM:SS" pill. Shipped 2026-05-18.
 - [32 — Audiobook export](32-audiobook-export.md) — Sideload to PocketBook Reader (Phase A: MP3.ZIP) via LAN download or sync folder; per-chapter ID3v2.4 tags, no re-encode, atomic writes.
 - [34 — MP3-folder export](34-mp3-folder-export.md) — Per-chapter MP3s in a sub-folder for folder-scanning audiobook apps (Smart AudioBook Player, BookPlayer, Audiobookshelf). Sync-folder destination only; APIC cover travels with each chapter.
+- [60 — Listen view decomposition](archive/60-listen-view-decompose.md) — Behaviour-neutral lift: `src/views/listen.tsx` (1136 → 319 lines) becomes a slice-wiring orchestrator that composes three new region sub-components under `src/components/listen/` (header, player+markers, downloads+queue). Zero spec modifications. Shipped 2026-05-19.
 
 ### I. Revisions & drift
 
@@ -157,3 +158,4 @@ breadcrumb so cross-references still resolve.
 - [57 — Listen-view download tiles](archive/57-download-tiles.md) — Wires the M4B chaptered + new MP3 ZIP tiles on the listen view to open `ExportAudiobookModal` with the format pre-filled; streaming-link tile remains "Coming soon" pending a slugged URL endpoint. Shipped 2026-05-19.
 - [58 — E2E coverage refresh](archive/58-e2e-coverage-refresh.md) — Un-quarantines `listen-playback` + `new-book-flow`; adds `binary-upload.spec.ts` covering EPUB / PDF / MOBI / AZW3 routing; applies file-level serial mode to five contention-flaky spec files. From 3-5 hard failures per `verify` run to 0. Shipped 2026-05-19.
 - [59 — Parallel Claude Code sessions](archive/59-parallel-claude-sessions.md) — `scripts/wt-new.mjs` spawns a git worktree on a new branch with non-colliding dev-server ports so multiple top-level `claude` sessions can run in parallel without fighting over `:5173` / `:8080` / `:9000` / `:5174`. Slot N → ports offset by `N*10`. `scripts/wt-list.mjs` lists active worktrees + their assignments. `vite.config.ts` + `playwright.config.ts` now env-driven; stock defaults preserved for the main worktree. Shipped 2026-05-19.
+- [60 — Listen view decomposition](archive/60-listen-view-decompose.md) — Behaviour-neutral lift of `src/views/listen.tsx` (1136 → 319 lines) into a slice-wiring orchestrator that composes three region sub-components under `src/components/listen/` (header, player+markers, downloads+queue). Pure-deletion diff (52 ins / 869 del); zero spec modifications. Gates Wave 3 of the v1.4.0 slate. Shipped 2026-05-19.
